+++ test/prepare.js
@@ -1,17 +1,15 @@
 'use strict';
 
-var glob = require('glob');
-var fs = require('fs');
-var path = require('path');
-var buffered = require('./util/buffered');
+const fs = require('fs');
+const path = require('path');
+const glob = require('glob');
 
+// Prepare fixtures
+const fixturesDir = __dirname + '/fixtures';
 
-// Preare fixtures
-var fixturesDir = __dirname + '/fixtures';
-
-glob.sync('prepare_*', { cwd: __dirname + '/fixtures' }).forEach(function (file) {
-    var contents = fs.readFileSync(fixturesDir + '/' + file);
-    var finalFile = file.replace(/^prepare_/, '').replace(/\.sh$/, '');
+glob.sync('prepare_*', { cwd: __dirname + '/fixtures' }).forEach((file) => {
+    const contents = fs.readFileSync(fixturesDir + '/' + file);
+    const finalFile = file.replace(/^prepare_/, '').replace(/\.sh$/, '');
 
     fs.writeFileSync(fixturesDir + '/' + finalFile, contents);
     fs.chmodSync(fixturesDir + '/' + finalFile, parseInt('0777', 8));
@@ -19,24 +17,10 @@ glob.sync('prepare_*', { cwd: __dirname + '/fixtures' }).forEach(function (file)
     process.stdout.write('Copied "' + file + '" to "' + finalFile + '"\n');
 });
 
-// Install spawn-sync for older node versions
-if (/^v0\.10\./.test(process.version)) {
-    process.stdout.write('Installing spawn-sync..\n');
-    buffered('spawn', 'npm', ['install', 'spawn-sync'], { stdio: 'inherit' }, function (err) {
-        if (err) {
-            throw err;
-        }
-
-        process.exit();
-    });
-}
-
 // Fix AppVeyor tests because Git bin folder is in PATH and it has a "echo" program there
 if (process.env.APPVEYOR) {
     process.env.PATH = process.env.PATH
     .split(path.delimiter)
-    .filter(function (entry) {
-        return !/\\git\\bin$/i.test(path.normalize(entry));
-    })
+    .filter((entry) => !/\\git\\bin$/i.test(path.normalize(entry)))
     .join(path.delimiter);
 }